// src/models/entreprise.ts
export class Entreprise {

  uuid: string ="";
    nif: string ="";
    utilisateur_uuid: string ="";
    utilisateur: string ="";

  nifNom: string ="";
     nomEntreprise: string ="";
    sigle: string ="";
    adresse: string ="";
    telephone: string ="";
      solde: string ="";
      
   
}